shapefile: cleanup
