+++ modules/shapefile/package.json
@@ -34,6 +34,7 @@
   },
   "dependencies": {
     "@loaders.gl/gis": "2.3.0-alpha.11",
-    "@loaders.gl/loader-utils": "2.3.0-alpha.11"
+    "@loaders.gl/loader-utils": "2.3.0-alpha.11",
+    "@loaders.gl/tables": "2.3.0-alpha.11"
   }
 }
