+++ modules/shapefile/test/index.js
@@ -1,7 +1,5 @@
 import './streaming/binary-chunk-reader.spec';
 
-import './parsers/parse-shp.spec';
-
 import './shp-loader.spec';
 import './dbf-loader.spec';
 import './shapefile-loader.spec';
